requirements_sheets.txt
# Append these lines to your requirements.txt
gspread==5.12.0
google-auth==2.34.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread-dataframe==3.3.1
